+++ src/console-app.c
@@ -94,7 +94,7 @@ int32_t main(int32_t argc, char *argv[]) {
             struct wsi_data *wsi_data = get_wsi_data(filename);
             // if format is supported and valid
             if (wsi_data->metadata_attributes != NULL) {
-                // TODO: print out the rest of information (label and macro dims if available
+                // TODO: print out the rest of information (label and macro dims if available)
                 print_metadata(wsi_data);
             }
             // handles invalid or unsupported file formats
@@ -102,7 +102,7 @@ int32_t main(int32_t argc, char *argv[]) {
                 fprintf(stderr, "Error: %s format\n", VENDOR_AND_FORMAT_STRINGS[wsi_data->format]);
                 exit(EXIT_FAILURE);
             }
-            free(wsi_data);
+            free_wsi_data(wsi_data);
         } else {
             fprintf(stderr, "No filename to check for vendor selected.\n");
             exit(EXIT_FAILURE);