+++ src/philips-based-io.h
@@ -8,7 +8,7 @@
 
 char *wipe_section_of_attribute(char *buffer, char *attribute);
 
-const char *get_value_from_attribute(char *buffer, char *attribute);
+char *get_value_from_attribute(char *buffer, char *attribute);
 
 char *anonymize_value_of_attribute(char *buffer, char *attribute);