+++ src/utils.h
@@ -24,7 +24,7 @@ void replace_str_inplace(char *original_str, const char *replace_str, const char
 
 bool starts_with(const char *str, const char *pre);
 
-const char *get_string_between_delimiters(const char *buffer, const char *delimiter1, const char *delimiter2);
+char *get_string_between_delimiters(const char *buffer, const char *delimiter1, const char *delimiter2);
 
 void remove_leading_spaces(char *str);